refactor: AdBridgeDispatcher
